Fixes bug in the way shares are saved over time
Fixes #81

debug_scripts/blocks_with_shares.js - Displays a list of blocks in your system + the number of shares in each.  Neat for debug purposes.

Due to the way that nodejs-pool saves shares, entire blocks of shares must be purged at once.  Currently, longRunner calls the cleanShareDB, which uses the top of the stack as the starting point for the scanning, and has probable issues with the way that shares are stored in blocks.  This was noted on a test pool as well, when the blocks were found too rapidly, and not enough share depth could be found after the purging.

This commit implements two new functions for this, cleanShareDBNew and getOldestLockedBlock, which are implemented in longRunnerNew to ensure this stops breaking.

getOldestLockedBlock returns the decompressed proto for the oldest block in the system that is unlocked === false, aka Locked.  This will ensure that the oldest blocks that are locked can be handled directly.

cleanShareDBNew implements a new scanning routine based on the following params:
If there is no locked block in the system, and PPS is enabled, start scanning from the current height - blocksRequired/2, as PPS blocks are added to the system after this point.
If there is no locked block in the system, and PPS is disabled, start scanning from the current height.

If there is a locked block, and PPS is enabled and the oldestLockedBlock.height > the current height - the blocksRequired/2, then use the current height - blocksRequired/2
If there is a locked block, and PPS is disabled or the oldestLockedBlock.height < the current height - the blocksRequired/2, then use the oldestLockedBlock.height as the start.

System also doesn't need to scan against the blocks to determine if they're unlocked already, because we know they are!  Well played!

This functionality is enabled with --module=longRunnerNew