Removing minerList from worker.  This seems small, but due to the way that JS handles arrays versus object lookups...
Long story short, doing in and for loops is about 9x faster than doing forEach and indexOf.  Based on this, we're doing first-round lookups and hashing.
Adding optimizations to avoid hitting LMDB unless it's actually needed.  How do we do that?  Well, we're magic of course!  The correct answer is that as we're loading objects, we now rely on LMDB to hold objects long-term, but use an in-memory cache (local_miner_cache) to store the last known .hash value for the miner.  Why is this important you ask?  I'm glad you asked!

Lets touch on what's really changing:
minerList is being removed, JS arrays are incredibly slow to index, and take extensive time to deal with.  On 47k objects scanning 47k objects, this is approximately 11 seconds.  Assuming that you have 47k objects and you're scanning a null list, in and indexOf are the same speed, about 3.2 seconds for 47k objects.  However, if you have 47k objects, and are scanning an object using the in statement, you're looking at 4 ms vs 11 seconds.  This will scale and change slightly with miner/identifier counts, however, there's other places we were doing this, with minerList versus globalMinerList.  GML is the more concerning of the two, but we got rid of a data object as well (minerList).

We now safely transmute the Array state of GML into GMO, or load GML as GMO if we detect that the types are correct.  As minerList is gone, we now rely on localStats.miners, which contains all of the local state information about miners, as an Object.  How handy right?  This lets us do in statements against it cleanly.  The second speedup is the LMC, which means we reduce our null hits to LMDB.  While not massive, a 47k database, with 30k inactive miners, will spend on average, .2ms per miner trying to load this effectively useless data, where we get free-ish in-memory lookups if the LMC contains the miner.  Initial startup will heat up the cache by loading the LMDB records once.  As 0 are no-ops for us for these functions, if the miner's state is 0, we'll never reload them, however, if they start mining again, they'll be in localStats.miners again, giving us access to their data and heating that cache back up again.  If they drop off, we'll blast them by setting their .hash back to 0, and dropping the state.

As we're doing this out of an in-memory cache, but we need to ensure the cache system in LMDB gets full docs, if we detect that we're loading these from the IMC, we set loaded_from_cache to true, in which case, we reload minerStats from the getCache and cacheDB, ensuring we have clean data.