Disabling ZMQ/Socket.IO until further notice.  Due to the design of the database and related data, it's not particularly easy to get data from these without really working fairly closely with MySQL, and query-bashing is a thing.
Disabling the sticky-sessions system as well.  As Socket.IO is down, this isn't needed.  The bigger issue is that the router ate/munged all the connection onto a single server rather than LB'ing them, because they were all proxied.  That product works well in it's handling of direct connections, but not proxied through something like Caddy or Nginx.  This should help even out load/balance.